Reporter: collapse ReportPrint inheritance; rename printAnalysis -> printReport
ReportPrint was a "for class organization only" parent class whose
methods were delegated to one-line via Reporter's public methods.
Merging it into Reporter eliminates the indirection and the
abstract-class artifact.

- All content of ReportPrint moved into Reporter: the three printing
  entry points (now public, as they should always have been), the
  private addFirstMainSection / addMainSection / printList helpers,
  the calculateHasFiftyMoveRule helper, and the two threshold
  constants.
- Reporter is now public final with a private constructor — pure
  utility class, no instantiation, no further subclassing.
- ReportPrint.java deleted.
- printAnalysis(...) renamed to printReport(...) everywhere it appears
  on Reporter and at all call sites (ReadMeTest, PrintSinglePgnAnalysis
  including its local helper). The class is "Reporter" producing
  "Report" output — the method name now matches.

calculateAnalysis(...) was left as-is; the user's instruction was
specifically about printAnalysis. If consistency with the renamed
package is wanted, calculateAnalysis -> calculateReport is a
follow-up rename for a separate pass.

950/950 tests still green.
