+++ src/test/java/com/dlb/chess/test/pgn/analysis/PrintSinglePgnAnalysis.java
@@ -7,7 +7,7 @@ import com.dlb.chess.test.pgntest.enums.PgnTest;
 public class PrintSinglePgnAnalysis {
 
   public static void main(String[] args) throws Exception {
-    printAnalysis("05_claim_for_own_move_correct_but_makes_move_on_board.pgn");
+    printReport("05_claim_for_own_move_correct_but_makes_move_on_board.pgn");
 
     final var pgnString = """
         [Event "Rated Blitz game"]
@@ -31,13 +31,13 @@ public class PrintSinglePgnAnalysis {
 
             1. c4 c5 2. Nc3 Nc6 3. Nf3 Nf6 { A35 English Opening: Symmetrical Variation, Four Knights Variation } 4. d4 cxd4 5. Nxd4 g6 6. e4 Bg7 7. Be3 O-O 8. Be2 d6 9. O-O Bd7 10. f3 Rc8 11. Qd2 Na5 12. b3 Ne8 13. Rac1 Nc7 14. Rfd1 Ne6 15. Qe1 Qb6 16. Qf2 Nc5 17. Nd5 Qd8 18. b4 Na4 19. bxa5 e6 20. Nc3 Qxa5 21. Nxa4 Bxa4 22. Rd2 Rc5 23. Nb3 Bxb3 24. axb3 Rc6 25. Rcd1 Qc7 26. Bd4 Bh6 27. Be3 Bg7 28. f4 e5 29. f5 f6 30. g4 Rd8 31. h4 Qa5 32. Bf3 Qb4 33. Rd3 b6 34. Bd2 Qa3 35. Be3 Qb4 36. Qd2 Qa3 37. Kg2 a6 38. Qf2 b5 39. cxb5 axb5 40. R1d2 Qb4 41. Kg3 Rc1 42. Rd1 Rc6 43. R1d2 Rc1 44. Rd1 Rc6 45. Bd2 Qa3 46. Be3 Qb4 47. Qd2 Qa3 48. Rc1 Rb6 49. Qc2 Ra6 50. Rc3 Rda8 51. Rc8+ Rxc8 52. Qxc8+ Bf8 53. Rc7 Ra8 54. Qd7 { Black resigns. } 1-0
         """;
-    Reporter.printAnalysis(pgnString);
+    Reporter.printReport(pgnString);
   }
 
-  static void printAnalysis(String pgnFileName) throws Exception {
+  static void printReport(String pgnFileName) throws Exception {
     final PgnTest pgnTest = PgnExpectedValue.findPgnTestPgnNotListed(pgnFileName);
     System.out.println(pgnFileName);
-    Reporter.printAnalysis(pgnTest.getFolderPath(), pgnFileName);
+    Reporter.printReport(pgnTest.getFolderPath(), pgnFileName);
   }
 
 }
